+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 <!-- FINALSPLEVEL keyword? -->
  <!-- add documentation: SCRAMBLECHECK kw, fsm>, neb>, rdkit_search>, standalone optimizer, non-inline constraints-->
 
+
+## FIRECODE 1.7.0 🔥 (WIP)
+- Restructured and expanded solvent module.
+- Implemented vibrational analysis via ASE including quasi-RRHO thermochemistry (as in GoodVibes) to get free energies.
+- Added FREQ keyword, that computes free energies at the end of the geometry optimization cycles.
+- Added T (temperature, in K), T_C (temperature, in °C), P (pressure, in atm), and C (concentration, in mol/L) keywords.
+<!-- add number of active constraints printout for parallel multithread functions -->
+
 ## FIRECODE 1.6.0 🔥 (March 11 2026)
 - Refreshed constraints handling in operators.
 - Added non-inline constraints specification after a molecule line. Syntax: "B/A/D i1 i2 [i3] [i4] [auto/value]"